+++ luasrc/model/cbi/snmpd-tn/snmpd.lua
@@ -3,9 +3,10 @@
 -- Anton Kikin <a.kikin@tano-systems.com>
 --
 
-local sys  = require "luci.sys"
-local util = require "luci.util"
-local uci  = require("luci.model.uci").cursor()
+local sys   = require "luci.sys"
+local util  = require "luci.util"
+local uci   = require("luci.model.uci").cursor()
+local snmpd = require("luci.tools.snmpd-tn")
 
 local m, s, o
 
@@ -200,68 +201,11 @@ local ro_community_src
 local rw_community
 local rw_community_src
 
-local function snmpd_get_community(self, section, secname)
-	if self.tag_error[section] then
-		com = self:formvalue(section)
-	else
-		com = self.map:get(section, self.alias or self.option)
-		if not com then
-			m.uci:foreach("snmpd", "com2sec", function(s)
-				if s.secname == secname then
-					com = s.community
-					return false
-				end
-			end)
-		end
-
-		if not com then
-			m.uci:foreach("snmpd", "com2sec6", function(s)
-				if s.secname == secname then
-					com = s.community
-					return false
-				end
-			end)
-		end
-
-		if not com then
-			com = self.default
-		end
-	end
-	return com
-end
-
-local function snmpd_get_community_src(self, section, secname)
-	local src = ""
-	if self.tag_error[section] then
-		src = self:formvalue(section)
-	else
-		src = self.map:get(section, self.alias or self.option)
-		if not src then
-			local com
-			if secname == "ro" then
-				com = ro_community
-			else
-				com = rw_community
-			end
-			src = m.uci:get("snmpd", com:cfgvalue(section), "source")
-			if not src then
-				src = self.default
-			end
-		end
-	end
-
-	return src
-end
-
 -- SNMPv1/SNMPv2c Read only community
 ro_community = s:taboption("v1v2c", Value, "ro_community",
 	translate("Read community"))
 ro_community.default = "public"
 ro_community.rmempty = false
-
-function ro_community.cfgvalue(self, section)
-	return snmpd_get_community(self, section, "ro")
-end
 
 ro_community_src = s:taboption("v1v2c", Value, "ro_community_src",
 	translate("Read community source"),
@@ -272,8 +216,13 @@ ro_community_src.default = "default"
 ro_community_src.rmempty = false
 ro_community_src.datatype = "or(host(0),ipmask)"
 
+function ro_community.cfgvalue(self, section)
+	return snmpd.community_cfgvalue(self, section, "ro")
+end
+
 function ro_community_src.cfgvalue(self, section)
-	return snmpd_get_community_src(self, section, "ro")
+	return snmpd.community_src_cfgvalue(self, section, "ro",
+		ro_community:cfgvalue(section))
 end
 
 -- SNMPv1/SNMPv2c Read/write community
@@ -281,10 +230,6 @@ rw_community = s:taboption("v1v2c", Value, "rw_community",
 	translate("Write community"))
 rw_community.default = "private"
 rw_community.rmempty = false
-
-function rw_community.cfgvalue(self, section)
-	return snmpd_get_community(self, section, "rw")
-end
 
 rw_community_src = s:taboption("v1v2c", Value, "rw_community_src",
 	translate("Write community source"),
@@ -295,8 +240,13 @@ rw_community_src.default = "localhost"
 rw_community_src.rmempty = false
 rw_community_src.datatype = "or(host(0),ipmask)"
 
+function rw_community.cfgvalue(self, section)
+	return snmpd.community_cfgvalue(self, section, "rw")
+end
+
 function rw_community_src.cfgvalue(self, section)
-	return snmpd_get_community_src(self, section, "rw")
+	return snmpd.community_src_cfgvalue(self, section, "rw",
+		rw_community:cfgvalue(section))
 end
 
 -----------------------------------------------------------------------------------
@@ -393,217 +343,27 @@ trap_community.rmempty = false
 
 -----------------------------------------------------------------------------------
 
-local function snmpd_community_delete(self, community)
-	if not community or community == "" then return end
-
-	-- delete sections
-	self.map:del(community)               -- com2sec
-	self.map:del(community .. "6")        -- com2sec6
-	self.map:del(community .. "_v1")      -- group
-	self.map:del(community .. "_v2c")     -- group
-	self.map:del(community .. "_usm")     -- group
-	self.map:del(community .. "_access")  -- access
-end
-
-local function snmpd_community_update(community, access, src, snmp_version, ip_protocol)
-	local secname = access
-	local sname
-
-	if snmp_version:match("v1%/v2c") then
-		-- com2sec section
-		if ip_protocol:match("ipv4") then
-			sname = community
-			if not uci:get("snmpd", sname) then
-				uci:set("snmpd", sname, "com2sec")
-			end
-			uci:set("snmpd", sname, "secname", secname)
-			uci:set("snmpd", sname, "source", src)
-			uci:set("snmpd", sname, "community", community)
-		end
-
-		-- com2sec6 section
-		if ip_protocol:match("ipv6") then
-			sname = community .. "6"
-			if not uci:get("snmpd", sname) then
-				uci:set("snmpd", sname, "com2sec6")
-			end
-			uci:set("snmpd", sname, "secname", secname)
-			uci:set("snmpd", sname, "source", src)
-			uci:set("snmpd", sname, "community", community)
-		end
-	end
-
-	-- groups
-	if snmp_version:match("v1%/v2c") then
-		sname = community .. "_v1"
-		if not uci:get("snmpd", sname) then
-			uci:set("snmpd", sname, "group")
-		end
-		uci:set("snmpd", sname, "group", community)
-		uci:set("snmpd", sname, "version", "v1")
-		uci:set("snmpd", sname, "secname", secname)
-
-		sname = community .. "_v2c"
-		if not uci:get("snmpd", sname) then
-			uci:set("snmpd", sname, "group")
-		end
-		uci:set("snmpd", sname, "group", community)
-		uci:set("snmpd", sname, "version", "v2c")
-		uci:set("snmpd", sname, "secname", secname)
-	end
-
-	if snmp_version:match("v3") then
-		sname = community .. "_usm"
-		if not uci:get("snmpd", sname) then
-			uci:set("snmpd", sname, "group")
-		end
-		uci:set("snmpd", sname, "group", community)
-		uci:set("snmpd", sname, "version", "usm")
-		uci:set("snmpd", sname, "secname", secname)
-	end
-
-	-- access
-	sname = community .. "_access"
-	if not uci:get("snmpd", sname) then
-		uci:set("snmpd", sname, "access")
-	end
-	uci:set("snmpd", sname, "group", community)
-	uci:set("snmpd", sname, "context", "none")
-	uci:set("snmpd", sname, "version", "any")
-	uci:set("snmpd", sname, "level", "noauth")
-	uci:set("snmpd", sname, "prefix", "exact")
-
-	if access == "ro" then
-		uci:set("snmpd", sname, "read", "all")
-		uci:set("snmpd", sname, "write", "none")
-		uci:set("snmpd", sname, "notify", "none")
-	else
-		uci:set("snmpd", sname, "read", "all")
-		uci:set("snmpd", sname, "write", "all")
-		uci:set("snmpd", sname, "notify", "all")
-	end
-end
-
 function snmp_enable.write(self, section, value)
-	local ro_update = false
-	local rw_update = false
-
-	local ro_community_prev = ""
-	local ro_community_new = ro_community:formvalue(section)
-
-	local ro_community_src_prev = ""
-	local ro_community_src_new = ro_community_src:formvalue(section)
-
-	local rw_community_prev = ""
-	local rw_community_new = rw_community:formvalue(section)
-
-	local rw_community_src_prev = ""
-	local rw_community_src_new = rw_community_src:formvalue(section)
-
-	local snmp_version_prev = ""
-	local snmp_version_new = snmp_version:formvalue(section)
-
-	local ip_protocol_prev = ""
-	local ip_protocol_new = ip_protocol:formvalue(section)
-
-	if ro_community_new == "" or rw_community_new == "" then
-		return
+	if snmpd.communities_config({
+		ro_community     = ro_community:formvalue(section),
+		ro_community_src = ro_community_src:formvalue(section),
+		rw_community     = rw_community:formvalue(section),
+		rw_community_src = rw_community_src:formvalue(section),
+		snmp_version     = snmp_version:formvalue(section),
+		ip_protocol      = ip_protocol:formvalue(section)
+	}) == true then
+		Flag.write(self, section, value)
 	end
-
-	--------
-
-	snmp_version_prev = uci:get("snmpd", "general", "snmp_version")
-	ip_protocol_prev = uci:get("snmpd", "general", "ip_protocol")
-
-	local com = "com2sec"
-	if ip_protocol_prev and ip_protocol_prev:match("ipv6") then com = "com2sec6" end
-	uci:foreach("snmpd", com, function(s)
-		if s.secname == "ro" then
-			ro_community_prev = s.community
-			ro_community_src_prev = s.source
-		end
-		if s.secname == "rw" then
-			rw_community_prev = s.community
-			rw_community_src_prev = s.source
-		end
-	end)
-
-	--------
-
-	if ro_community_src_new ~= ro_community_src_prev then ro_update = true end
-	if rw_community_src_new ~= rw_community_src_prev then rw_update = true end
-
-	if ro_community_new ~= ro_community_prev then
-		snmpd_community_delete(self, ro_community_prev)
-		ro_update = true
-	end
-
-	if rw_community_new ~= rw_community_prev then
-		snmpd_community_delete(self, rw_community_prev)
-		rw_update = true
-	end
-
-	if snmp_version_new ~= snmp_version_prev or
-	   ip_protocol_new ~= ip_protocol_prev then
-		ro_update = true
-		rw_update = true
-	end
-
-	if ro_update then
-		snmpd_community_update(ro_community_new, "ro",
-			ro_community_src_new, snmp_version_new, ip_protocol_new)
-	end
-
-	if rw_update then
-		snmpd_community_update(rw_community_new, "rw",
-			rw_community_src_new, snmp_version_new, ip_protocol_new)
-	end
-
-	Flag.write(self, section, value)
 end
 
 function trap_enable.write(self, section, value)
 	local t_version = trap_snmp_version:formvalue(section)
 	local t_enable = value
-
 	local t_host = trap_host:formvalue(section)
 	local t_port = trap_port:formvalue(section)
 	local t_community = trap_community:formvalue(section)
 
-	local traps
-	
-	traps = 0
-	m.uci:foreach("snmpd", "trapsink", function(s)
-		traps = traps + 1
-		if traps > 1 or t_enable == "0" or t_version ~= "v1" then
-			self.map:del(s[".name"])
-		end
-	end)
-
-	traps = 0
-	m.uci:foreach("snmpd", "trap2sink", function(s)
-		traps = traps + 1
-		if traps > 1 or t_enable == "0" or t_version ~= "v2c" then
-			self.map:del(s[".name"])
-		end
-	end)
-
-	if t_enable == "1" then
-		local sink
-		if t_version == "v1" then
-			sink = "trapsink"
-		else
-			sink = "trap2sink"
-		end
-
-		local s = m.uci:get_first("snmpd", sink)
-		if not s then
-			s = self.map:add(sink)
-		end
-		self.map:set(s, "community", t_community)
-		self.map:set(s, "host", t_host)
-		self.map:set(s, "port", t_port)
-	end
+	snmpd.traps_config(t_enable, t_version, t_host, t_port, t_community)
 
 	Flag.write(self, section, value)
 end