+++ CHANGELOG.md
@@ -6,8 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 All dates in this document are in `DD.MM.YYYY` format.
 
+## [Unreleased]
+
+### Changed
+- Moved some functions from CBI model code to module 'luci.tools.snmpd-tn'.
+
 ## [Version 0.9.0] (22.01.2019)
 
 Initial release
 
+[Unreleased]: https://github.com/tano-systems/luci-app-snmpd-tn/tree/master
 [Version 0.9.0]: https://github.com/tano-systems/luci-app-snmpd-tn/releases/tag/v0.9.0